djangomission/requirements.txt
asgiref==3.2.10
Django==3.1.14
django-environ==0.9.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
PyJWT==2.5.0
pytz==2022.2.1
sqlparse==0.4.2
